# bytes, strings: micro-optimize EqualFold