--- a/src/strings/strings.go
+++ b/src/strings/strings.go
@@ -1194,7 +1194,7 @@ func ReplaceAll(s, old, new string) string {
 func EqualFold(s, t string) bool {
 	// ASCII fast path
 	i := 0
-	for ; i < len(s) && i < len(t); i++ {
+	for n := min(len(s), len(t)); i < n; i++ {
 		sr := s[i]
 		tr := t[i]
 		if sr|tr >= utf8.RuneSelf {